Fix bugs when handling same day and bed and breakfast rules
In ApplySameDayMatchingRule DateTime is compared for equality instead of Datetime.Date, this results in trade in the same day never trigger the rule unless the trade is at the same time.
In ApplyBedAndBreakfastMathingRule DateTime is compared for equality, causing a sell trade that occurs early in the morning and a buy trade that happen exactly 30 days later but in the afternoon not to trigger the rule. The rules should count for calander days.