Correct minor formating issue.
1.) In a line such as
"Bed and breakfast: 150 units of the acquisition trade against 150 units of the disposal trade. acquisition cost is"
the start of the new sentence should be capitalised, i.e. "Acquisition cost" rather than "acquisition cost".

2.) The summary explanation that you have implemented for each complex disposal as I suggested, i.e. lines like
"Resulting overall gain for this disposal: X + Y = Z"
are great but there should be an extra new line (empty line) after that before the next block (which starts with "********")